feat(combo-button): move combobuttonitem, remove mocks file
