--- src/components/ComboButton/_mocks_/index.js
@@ -1,15 +1,0 @@
-/**
- * @file Combo button mocks.
- * @copyright IBM Security 2019
- */
-
-import { ArrowRight20 } from '@carbon/icons-react';
-
-const primaryAction = {
-  label: 'Start a task',
-  iconDescription: 'arrow right',
-  renderIcon: ArrowRight20,
-  onClick: () => {},
-};
-
-export default primaryAction;
