+++ src/index.js
@@ -8,8 +8,7 @@ export ErrorPage from './components/ErrorPage';
 export { Card, CardSkeleton } from './components/Card';
 export CodeSnippet, { CodeSnippetSkeleton } from './components/CodeSnippet';
 export ComboBox from './components/ComboBox';
-export ComboButton from './components/ComboButton';
-export ComboButtonItem from './components/ComboButtonItem';
+export ComboButton, { ComboButtonItem } from './components/ComboButton';
 export ContentSwitcher from './components/ContentSwitcher';
 export CopyButton from './components/CopyButton';
 export DataDecorator, { Decorator } from './components/DataDecorator';
